+++ zip/Zip.php
@@ -1,9 +1,9 @@
 <?php
 /*
- |  FoxCMS      Content Management Simplified <www.foxcms.org>
+ |  A ZipArchive and PKZIP PHP helper class.
  |  @file       ./includes/ZIP.php
  |  @author     SamBrishes@pytesNET
- |  @version    0.2.0 [0.2.0] - Beta
+ |  @version    0.2.1 [0.2.1] - Beta
  |
  |  @license    GNU GPL v3
  |  @copyright  Copyright © 2015 - 2018 SamBrishes, pytesNET <pytes@gmx.net>
@@ -12,9 +12,9 @@
  |              Copyright © 2008 - 2009 Philippe Archambault <philippe.archambault@gmail.com>
  */
 /*
- |  The following websites contains all required informations, which were unavoidable for the 
+ |  The following websites contains all required informations, which were unavoidable for the
  |  creation of this class:
- |  
+ |
  |  -   https://pkware.cachefly.net/webdocs/casestudies/APPNOTE.TXT
  |  -   https://users.cs.jmu.edu/buchhofp/forensics/formats/pkzip.html
  |  -   https://php.net/manual/class.ziparchive.php
@@ -25,19 +25,19 @@
         const VERSION = "\x14\x00";
         const SIGNATURE = "\x50\x4b";
         const COMPRESSION = "\x08\x00";
-        
+
         /*
          |  SETTINGs
          */
         public $zipArchive = false;
         public $compression = 6;
-        
+
         /*
          |  ZIP ARCHIVE
          */
         private $zipFilename;
         private $zipInstance;
-        
+
         /*
          |  FALLBACK
          */
@@ -45,7 +45,7 @@
         private $headers = array();
         private $central = array();
         private $counter = 0;
-        
+
         /*
          |  CONSTRUCTOR
          |  @since  0.2.0
@@ -65,7 +65,7 @@
             }
             $this->compression = ($compression >= -1 && $compression <= 9)? $compression: 6;
         }
-        
+
         /*
          |  DESTRUCTOR
          |  @since  0.2.0
@@ -73,56 +73,59 @@
         public function __destruct(){
             $this->clear(false);
         }
-        
+
         /*
          |  HELPER :: CONVERT UNIX TO DOS TIME
-         |  @since  0.2.0
+         |  @since  0.2.1
          |
          |  @param  int     The respective timestamp as INTEGER.
          */
-        private function hexTime($time){
+        protected function msDOSTime($time){
             $array = getdate((is_int($time) && $time > 0)? $time: time());
-            if($array["year"] < 1980){
+            if($array["year"] < 1980 || $array["year"] > 2107){
                 $array = getdate(time());
             }
-            
-            // Return as HEX
-            $time = dechex(
-                (($array["year"] - 1980) << 25) | ($array["mon"] << 21) | ($array["mday"] << 16) | 
-                ($array["hours"] << 11) | ($array["minutes"] << 5)| ($array["seconds"] >> 1)
+
+            // Return as DEC
+            return (
+                (($array["year"]-1980 << 25)) |
+                (($array["mon"]       << 21)) |
+                (($array["mday"]      << 16)) |
+                (($array["hours"]     << 11)) |
+                (($array["minutes"]   <<  5)) |
+                (($array["seconds"]   >>  1))
             );
-            $time = $time[6].$time[7].$time[4].$time[5].$time[2].$time[3].$time[0].$time[1];
-            return pack("H*", $time);
         }
-        
+
         /*
          |  ADD A FILE
          |  @since  0.1.0
-         |  @update 0.2.0
+         |  @update 0.2.1
          |
          |  @param  string  The relative or absolute filepath or the respective file content.
          |  @param  string  The local path within the archive file.
          |  @param  int     The timestamp to use.
-         |  
+         |  @param  string  The optional file comment or just an empty string.
+         |
          |  @return bool    TRUE on success, FALSE on failure.
          */
-        public function addFile($data, $path, $time = 0){
+        public function addFile($data, $path, $time = 0, $comment = ""){
             if((!is_string($data) && !is_numeric($data)) || !is_string($path)){
                 return false;
             }
-            
+
             // Sanitize Data
             if(is_string($data) && file_exists($data) && is_file($data)){
                 $data = file_get_contents($data);
             }
             $path = trim(str_replace("\\", "/", $path), "/");
-            $time = $this->hexTime($time);
-            
+            $time = $this->msDOSTime($time);
+
             // Zip Archive
             if($this->zipArchive){
                 return $this->zipInstance->addFromString($path, $data);
             }
-            
+
             // Fallback
             $crcval = crc32($data);
             $length = strlen($data);
@@ -131,9 +134,9 @@
             } else {
                 $gzcval = gzcompress($data, $this->compression);
             }
-            $gzcval = substr(substr($gzcval, 0, strlen($gzcval) - 4), 2); // Fix CRC Bug
+            $gzcval = substr($gzcval, 2, strlen($gzcval) - 6); // Fix CRC-32 Bug
             $gzclen = strlen($gzcval);
-            
+
             /*
              |  LOCAL FILE HEADER
              |  01      SIGNATURE
@@ -149,21 +152,20 @@
              |  11      The relative path / filename inside the archive.
              |  12      The main file data value.
              */
-            $this->headers[] = implode("", array(
-                self::SIGNATURE . "\x03\x04", 
-                self::VERSION, 
-                self::FLAGS, 
-                self::COMPRESSION,
-                $time,
-                pack("V", $crcval),
-                pack("V", $gzclen),
-                pack("V", $length),
-                pack("v", strlen($path)),
-                pack("v", 0),
-                $path,
-                $gzcval
-            ));
-            
+            $this->headers[] =
+                self::SIGNATURE . "\x03\x04" .
+                self::VERSION .
+                self::FLAGS .
+                self::COMPRESSION .
+                pack("V", $time) .
+                pack("V", $crcval) .
+                pack("V", $gzclen) .
+                pack("V", $length) .
+                pack("v", strlen($path)) .
+                pack("v", 0) .
+                $path .
+                $gzcval;
+
             /*
              |  CENTRAL DIRECTORY RECORD
              |  01      SIGNATURE
@@ -183,41 +185,45 @@
              |  15      External file attributes.
              |  16      Offset of the local file header.
              |  17      The relative path / filename inside the archive.
+             |  18      The file comment.
              */
-            $this->central[] = implode("", array(
-                self::SIGNATURE . "\x01\x02", 
-                "\x00\x00", 
-                self::VERSION, 
-                self::FLAGS, 
-                self::COMPRESSION,
-                $time,
-                pack("V", $crcval),
-                pack("V", $gzclen),
-                pack("V", $length),
-                pack("v", strlen($path)),
-                pack("v", 0),
-                pack("v", 0),
-                pack("v", 0),
-                pack("v", 0),
-                pack("V", 32),
-                pack("V", $this->offset),
-                $path
-            ));
-            
+            $this->central[] =
+                self::SIGNATURE . "\x01\x02" .
+                "\x00\x00" .
+                self::VERSION .
+                self::FLAGS .
+                self::COMPRESSION .
+                pack("V", $time) .
+                pack("V", $crcval) .
+                pack("V", $gzclen) .
+                pack("V", $length) .
+                pack("v", strlen($path)) .
+                pack("v", 0) .
+                pack("v", strlen($comment)) .
+                pack("v", 0) .
+                pack("v", 0) .
+                pack("V", 32) .
+                pack("V", $this->offset) .
+                $path .
+                $comment;
+
             // Count Offset and Return
             $this->offset += strlen($this->headers[count($this->headers)-1]);
             return true;
         }
-        
+
         /*
          |  ADD MULTIPLE FILES
          |  @since  0.2.0
+         |  @update 0.2.1
          |
          |  @param  array   Multiple 'local/file/path' => "filepath/or/filecontent" ARRAY pairs.
+         |  @param  int     The timestamp to use for all files.
+         |  @param  string  The optional comment or just an empty string for alle respective files.
          |
          |  @return int     The number of successfully added elements / files.
          */
-        public function addFiles($array){
+        public function addFiles($array, $time = 0, $comment = ""){
             if(!is_array($array)){
                 return false;
             }
@@ -226,7 +232,7 @@
             }
             return array_filter(array_values($array));
         }
-        
+
         /*
          |  ADD FOLDER
          |  @since  0.2.0
@@ -245,25 +251,25 @@
             if(!file_exists($path) || !is_dir($path)){
                 return false;
             }
-            
+
             // Chech Path
             $path = str_replace(array("/", "\\"), DIRECTORY_SEPARATOR, realpath($path));
             if(strpos($path, DIRECTORY_SEPARATOR) !== strlen($path)-1){
                 $path .= DIRECTORY_SEPARATOR;
             }
-            
+
             // Check Local
             if(!is_string($local)){
                 $local = "";
             }
             $local = trim(str_replace("\\", "/", $local), "/") . "/";
-            
+
             // Start Flow
             $this->counter = 0;
             $this->addFolderFlow($path, "", $local, !!$recursive, !!$empty);
             return $this->counter;
         }
-        
+
         /*
          |  HELPER :: ADD FOLDER LOOP
          |  @since  0.2.0
@@ -284,7 +290,7 @@
             if(!empty($path)){
                 $path .= DIRECTORY_SEPARATOR;
             }
-            
+
             $count = 0;
             $handle = opendir($base . $path);
             while(($file = readdir($handle)) !== false){
@@ -310,66 +316,68 @@
             closedir($handle);
             return $count;
         }
-        
+
         /*
          |  ADD EMPTY FOLDER
          |  @since  0.2.0
+         |  @update 0.2.1
          |
          |  @param  string  The local path structure within the zip file.
-         |  
+         |  @param  int     The timestamp to use.
+         |  @param  string  The optional file comment or just an empty string.
+         |
          |  @return bool    TRUE on success, FALSE on failure.
          */
-        public function addEmptyFolder($path){
+        public function addEmptyFolder($path, $time = 0, $comment = ""){
             $path = trim(str_replace("\\", "/", $path), "/") . "/";
-            $time = $this->hexTime(time());
-            
+            $time = $this->msDOSTime($time);
+
             // ZipArchive
             if($this->zipArchive){
                 return $this->zipInstance->addEmptyDir($path);
             }
-            
+
             // Add Header
-            $this->headers[] = implode("", array(
-                self::SIGNATURE . "\x03\x04", 
-                self::VERSION, 
-                self::FLAGS, 
-                "\x00\x00", 
-                $time,
-                pack("V", 0),
-                pack("V", 0),
-                pack("V", 0),
-                pack("v", strlen($path)),
-                pack("v", 0),
-                $path,
-                ""
-            ));
-            
+            $this->headers[] =
+                self::SIGNATURE . "\x03\x04" .
+                self::VERSION .
+                self::FLAGS .
+                "\x00\x00" .
+                pack("V", $time) .
+                pack("V", 0) .
+                pack("V", 0) .
+                pack("V", 0) .
+                pack("v", strlen($path)) .
+                pack("v", 0) .
+                $path .
+                "";
+
             // Add Central
-            $this->central[] = implode("", array(
-                self::SIGNATURE . "\x01\x02", 
-                "\x14\x03", 
-                self::VERSION, 
-                self::FLAGS, 
-                "\x00\x00", 
-                $time,
-                pack("V", 0),
-                pack("V", 0),
-                pack("V", 0),
-                pack("v", strlen($path)),
-                pack("v", 0),
-                pack("v", 0),
-                pack("v", 0),
-                pack("v", 0),
-                "\x00\x00\xFF\x41",
-                pack("V", $this->offset),
-                $path
-            ));
-            
+            $this->central[] =
+                self::SIGNATURE . "\x01\x02" .
+                "\x14\x03" .
+                self::VERSION .
+                self::FLAGS .
+                "\x00\x00" .
+                pack("V", $time) .
+                pack("V", 0) .
+                pack("V", 0) .
+                pack("V", 0) .
+                pack("v", strlen($path)) .
+                pack("v", 0) .
+                pack("v", strlen($comment)) .
+                pack("v", 0) .
+                pack("v", 0) .
+                "\x00\x00\xFF\x41" .
+                pack("V", $this->offset) .
+                $path .
+                $comment;
+
             // Count Offset and Return
             $this->offset += strlen($this->headers[count($this->headers)-1]);
             return true;
         }
-        
+
         /*
          |  CLEAR DATA STRINGs
          |  @since  0.1.0
@@ -394,27 +402,27 @@
             $this->central = array();
             return true;
         }
-        
+
         /*
          |  DUMBS OUT THE FILE
          |  @since  0.1.0
-         |  @update 0.2.0
+         |  @update 0.2.1
          */
         public function file(){
-            $comment = "PK-Zipped with https://www.github.com/SamBrishes/FoxCMS";
-            
+            $comment = "PKZipped with https://github.com/SamBrishes/FoxCMS/tree/helpers/zip";
+
             // ZipArchive
             if($this->zipArchive){
                 $this->zipInstance->setArchiveComment($comment);
                 $this->zipInstance->close();
-                
+
                 $content = file_get_contents($this->zipFilename);
-                
+
                 $this->zipInstance = new ZipArchive();
                 $this->zipInstance->open($this->zipFilename);
                 return $content;
             }
-            
+
             // Fallback
             $headers = implode("", $this->headers);
             $central = implode("", $this->central);
@@ -433,23 +441,20 @@
              |  11      The length of the following comment field.
              |  12      The archive comment.
              */
-            return implode("", array(
-                $headers,
-                $central,
-                self::SIGNATURE . "\x05\x06", 
-                "\x00",
-                "\x00",
-                "\x00",
-                "\x00",
-                pack("v", count($this->central)),
-                pack("v", count($this->central)),
-                pack("V", strlen($central)),
-                pack("V", strlen($headers)),
-                pack("v", strlen($comment)),
-                $comment
-            ));
+            return $headers . $central .
+                self::SIGNATURE . "\x05\x06" .
+                "\x00" .
+                "\x00" .
+                "\x00" .
+                "\x00" .
+                pack("v", count($this->central)) .
+                pack("v", count($this->central)) .
+                pack("V", strlen($central)) .
+                pack("V", strlen($headers)) .
+                pack("v", strlen($comment)) .
+                $comment;
         }
-        
+
         /*
          |  STORE THE ZIP FILE
          |  @since  0.1.0
@@ -464,7 +469,7 @@
             if(file_exists($filename) && !$overwrite){
                 return false;
             }
-            
+
             // Zip Archive
             if($this->zipArchive){
                 if(is_a($this->zipInstance, "ZipArchive")){
@@ -472,18 +477,18 @@
                 }
                 if(@file_put_contents($filename, file_get_contents($this->zipFilename))){
                     @unlink($this->zipFilename);
-                    
+
                     $this->zipFilename = $filename;
                     $this->zipInstance = new ZipArchive();
                     return $this->zipInstance->open($this->zipFilename);
                 }
                 return false;
             }
-            
+
             // Fallback
             return @file_put_contents($filename, $this->file()) !== false;
         }
-        
+
         /*
          |  DOWNLOAD THE FILE
          |  @since  0.1.0